AdventOfCode_Day7_BothStars

+++ source/AdventOfCode/PrintQueue/inc/PrintQueue.hpp
@@ -7,13 +7,13 @@
 
 class PrintQueue {
 public:
-  unsigned long run(const std::vector<std::string> &);
+  unsigned long long run(const std::vector<std::string> &);
 
 private:
   void processInput(const std::vector<std::string> &);
-  unsigned long getResultFromUpdates();
-  unsigned long getResultIfCorrect(std::vector<unsigned long>);
+  unsigned long long getResultFromUpdates();
+  unsigned long long getResultIfCorrect(std::vector<unsigned long long>);
 
-  std::map<unsigned long, std::set<unsigned long>> pagesAfterKeyPage;
-  std::vector<std::vector<unsigned long>> updates;
+  std::map<unsigned long long, std::set<unsigned long long>> pagesAfterKeyPage;
+  std::vector<std::vector<unsigned long long>> updates;
 };
