+++ source/AdventOfCode/RedNosedReports/inc/RedNosedReports.hpp
@@ -6,13 +6,13 @@
 
 class RedNosedReports {
 public:
-  unsigned long run(const std::vector<std::string> &);
+  unsigned long long run(const std::vector<std::string> &);
 
 private:
-  std::vector<int> getDifferences(const std::vector<unsigned long> &);
+  std::vector<int> getDifferences(const std::vector<unsigned long long> &);
   bool differenceValidator(const std::vector<int> &);
   void recheckOnMistake(std::function<bool(int)> isMistake,
                         const std::vector<int> &differences,
-                        const std::vector<unsigned long> &report,
-                        unsigned long &result);
+                        const std::vector<unsigned long long> &report,
+                        unsigned long long &result);
 };
