+++ source/AdventOfCode/Utils/inc/Utils/Parsers.hpp
@@ -6,5 +6,5 @@
 namespace parsers {
 constexpr auto digits{"0123456789"};
 
-std::vector<unsigned long> parseNumbersFromLine(std::string line);
+std::vector<unsigned long long> parseNumbersFromLine(std::string line);
 } // namespace parsers
